requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.46
PyMySQL==1.0.2
python-dotenv==0.19.0
Flask-Migrate==3.1.0
